{{cookiecutter.project_slug}}/requirements.txt
--index-url https://pip.antidot.net/root/unstable/+simple/

pytest
wheel
fluid-api
